pyproject.toml
[tool.poetry]
name = "liana"
version = "1.0.5"
description = "LIANA+: a one-stop-shop framework for cell-cell communication"
authors = ["Daniel Dimitrov <daniel.dimitrov@uni-heidelberg.de>"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Natural Language :: English",
]
license = "GPLv3"
readme = "README.md"
repository = "https://github.com/saezlab/liana-py"
homepage = "https://liana-py.readthedocs.io"
include = ["LICENSE"]

[tool.poetry.dependencies]
python = ">=3.8 <3.12"
anndata = ">=0.7.4"
mudata = "*"
scanpy = ">=1.8.0"
numba = ">=0.54.0"
plotnine = ">=0.10.0"
tqdm = "^4.0"
omnipath = {version = ">=1.0.6", optional = true}
decoupler = {version = ">=1.4.0", optional = true}
pydeseq2 = {version = ">=0.3.5", optional = true}
squidpy = {version = "*", optional = true}
cell2cell = {version = "*", optional = true}
muon = {version = "*", optional = true}
mofax = {version = "*", optional = true}
corneto = {version = "0.9.1-alpha.4", optional = true}
cvxpy = {version = "^1.2.1", optional = true}
cylp = {version = "^0.91.5", optional = true}
pre-commit = {version = ">=3.0.0", optional = true}
requests = {version = "^2.25.1", optional = true}
docrep = ">=0.3.1"

[tool.poetry.dev-dependencies]
pytest = ">=6.0"
pytest-cov = "*"
flake8 = "*"
sphinx = "^4.0"
nbsphinx = "^0.8"
sphinx_automodapi = "^0.14"
ipython = "^7.0"
ipykernel = "^6.0"
pandoc = "^1.0"
numpydoc = "^1.1"

[tool.poetry.extras]
common = ["decoupler", "omnipath", "requests"]
full = ["decoupler", "omnipath", "pydeseq2", "squidpy", "cell2cell", "muon", "mofax", "corneto", "cvxpy", "cylp", "requests"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
